api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python-headless
numpy
openai
python-dotenv